requirements.txt
# Tubi Content Trend Lifecycle Analyzer — Python dependencies
# Install with: pip install -r requirements.txt
# Tested on Python 3.10+ (including Python 3.13)

# --- Environment ---
python-dotenv>=1.0.0

# --- HTTP ---
requests>=2.31.0

# --- Data manipulation ---
pandas>=2.2.0
numpy>=2.0.0

# --- Survival analysis ---
lifelines>=0.29.0

# --- Time-series smoothing (inflection detection) ---
scipy>=1.14.0

# --- Machine learning utilities ---
scikit-learn>=1.5.0

# --- Visualization ---
matplotlib>=3.9.0
seaborn>=0.13.2

# --- Notebook ---
jupyter>=1.0.0
ipykernel>=6.29.0
